# [rlgl] Optimize rlReadScreenPixels